# api/openapi/xflow-v1.yaml
openapi: 3.1.0
info:
  title: xflow Workflow Management API
  version: 1.0.0
  description: |
    User-facing HTTP contract for xflow. This spec mirrors the routes that are
    actually registered on the apiserver mux (docs/design/API-SPECIFICATION.md
    §7, §10). The runner protocol face — including POST /v1/executions
    (entry-seed) and the entire /v1/runners/* family — is intentionally absent:
    per spec §0.1 / §10 the OpenAPI contract describes only the user face.

    Every user-face JSON response is wrapped in the Envelope (§3). The only
    exceptions are the raw-byte streams (GET /v1/supplies/{name},
    GET|HEAD /v1/artifacts/{digest}) whose 2xx bodies are octet-streams (their
    failure bodies are still JSON Envelopes — spec §3.4) and the /healthz and
    /readyz probes, which load balancers parse directly and must not be
    enveloped (§7).

    Delivery semantics, stated once here instead of per endpoint: the handler
    boundary and the Runner Protocol are at-least-once — lease replay, node
    execution timeout and failover can each invoke the same handler twice — so
    business side effects are the host's responsibility, with
    `execution_id` + `node_name` as the recommended idempotency key. Those are
    the faces this document deliberately omits (spec §0.1 / §10); see
    docs/design/RELEASE-GATES.md §3 and §3.1. No endpoint below promises
    exactly-once delivery.
  contact:
    name: xflow maintainers
    url: https://github.com/xbcio/xflow
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/workflows/execute:
    post:
      operationId: executeInlineWorkflow
      summary: Compile and execute an inline workflow definition
      description: |
        Merges the old submit/invoke shapes: when entry is empty the graph's
        default start node is used; when set, that entry node drives the
        explicit-entry path (spec §7, §9.1).
      tags: [Workflows]
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteWorkflowRequest"
      responses:
        "200":
          description: Execution started.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ExecuteWorkflowResponse"
        "400":
          description: |
            The definition is invalid. `workflow_compile_failed` (ErrorEnvelope)
            when the compiler rejects it; `workflow_param_invalid`
            (ParamInvalidErrorEnvelope, carrying `data.param_issues`) when the
            server runs ParamSpec validation in enforce mode and a node
            parameter violates its node type's Descriptor.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorEnvelope"
                  - $ref: "#/components/schemas/ParamInvalidErrorEnvelope"
              examples:
                workflowParamInvalid:
                  summary: Enforce-mode ParamSpec validation rejected a node parameter.
                  value:
                    success: false
                    code: workflow_param_invalid
                    message: 'workflow "orders": 1 invalid parameter(s): node "pause": duration is required'
                    data:
                      param_issues:
                        - node: pause
                          path: /parameters/duration
                          code: required
                          message: duration is required
                          severity: error
                    trace_id: 0123456789abcdef0123456789abcdef
        default:
          $ref: "#/components/responses/DefaultError"
...
  /v1/executions/{id}/signals:
    post:
      operationId: deliverSignal
      summary: Deliver a signal to an execution
      description: Delivers a named signal with an optional data payload.
      tags: [Executions]
      parameters:
        - $ref: "#/components/parameters/ExecutionId"
        - $ref: "#/components/parameters/XRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalRequest"
      responses:
        "200":
          description: Signal accepted.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AcceptedResponse"
        default:
          $ref: "#/components/responses/DefaultError"
  /v1/executions/{id}/signals/{name}:
    delete:
      operationId: revokeSignal
      summary: Revoke a signal
      description: Revokes a not-yet-consumed signal by name. The name travels in the path (§9.1).
      tags: [Executions]
      parameters:
        - $ref: "#/components/parameters/ExecutionId"
        - $ref: "#/components/parameters/SignalName"
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: Signal revoked.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RevokedResponse"
        default:
          $ref: "#/components/responses/DefaultError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Server-issued bearer token bound to a (subject, namespace, scopes) principal.
  parameters:
...
    ExecutionId:
      name: id
      in: path
      required: true
      description: Execution id.
      schema:
        type: string
    SignalName:
      name: name
      in: path
      required: true
      description: Signal name.
      schema:
        type: string
...
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: |
        Optional client-supplied request id (charset [A-Za-z0-9._-], length
        1..128). When valid it is echoed verbatim in the response header for
        client-side retry correlation (spec §5.2). Never appears in the body.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{1,128}$'
...
  headers:
    XRequestId:
      description: Echoed X-Request-Id (only when the inbound value was valid).
      schema:
        type: string
...
  responses:
    DefaultError:
      description: Error response — a failure Envelope with success=false and no data field.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Envelope:
      type: object
      description: |
        The user-face response envelope (spec §3). On success success=true,
        code="200" (string), and data carries the payload. On failure
        success=false, data is omitted (not null) — except for
        workflow_param_invalid, which carries data.param_issues — and code
        carries a stable snake_case business identifier. trace_id is server-authoritative
        (W3C 32-hex).
      required: [success, code, message, trace_id]
      properties:
        success:
          type: boolean
          description: Strictly consistent with the HTTP status (2xx → true).
        code:
          type: string
          description: Machine-readable identifier. "200" on success; stable snake_case business code on failure.
        message:
          type: string
          description: Human-readable text. "" on success. Never carries credentials, node output, SQL, or stack traces (§3.5).
        data:
          description: Payload on success. Omitted on failure, except workflow_param_invalid (see ParamInvalidErrorEnvelope).
        trace_id:
          type: string
          description: Server-authoritative W3C trace id (32 hex).
          pattern: '^[0-9a-f]{32}$'
    ErrorEnvelope:
      type: object
      description: The failure shape of the Envelope — data is absent on failure.
      additionalProperties: false
      required: [success, code, message, trace_id]
      properties:
        success:
          type: boolean
          description: Always false on this shape.
          example: false
        code:
          type: string
          description: Stable snake_case business error code.
          example: workflow_not_found
        message:
          type: string
          description: Human-readable text. May change; clients must branch on code.
          example: workflow not found
        trace_id:
          type: string
          description: Server-authoritative W3C trace id (32 hex).
          pattern: '^[0-9a-f]{32}$'
...
    ParamInvalidErrorEnvelope:
      type: object
      description: |
        The failure Envelope of `workflow_param_invalid` — the one failure that
        carries data, so a client can place every issue on its field.
      additionalProperties: false
      required: [success, code, message, data, trace_id]
      properties:
        success:
          type: boolean
          description: Always false on this shape.
          example: false
        code:
          type: string
          description: Always workflow_param_invalid on this shape.
          enum: [workflow_param_invalid]
        message:
          type: string
          description: Human-readable summary. May change; clients must branch on code.
        data:
          type: object
          description: The rejected definition's findings.
          additionalProperties: false
          required: [param_issues]
          properties:
            param_issues:
              type: array
              description: Every finding, warnings included.
              items:
                $ref: "#/components/schemas/ParamIssue"
        trace_id:
          type: string
          description: Server-authoritative W3C trace id (32 hex).
          pattern: '^[0-9a-f]{32}$'
    ExecuteWorkflowRequest:
      type: object
      required: [workflow]
      properties:
        workflow:
          $ref: "#/components/schemas/WorkflowDef"
        entry:
          type: string
          description: Explicit entry node; empty uses the graph default start node.
        input:
          type: object
          additionalProperties: true
        params:
          type: object
          additionalProperties: true
...
    ExecuteWorkflowResponse:
      type: object
      required: [execution_id]
      properties:
        execution_id:
          type: string
        param_issues:
          type: array
          items:
            $ref: "#/components/schemas/ParamIssue"
          description: |
            ParamSpec validation findings for an inline definition
            (POST /v1/workflows/execute only; a registered workflow was
            validated at registration). Same semantics as
            RegisterWorkflowResponse.param_issues.
    SignalRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        data:
          type: object
          additionalProperties: true
    AcceptedResponse:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: boolean
          example: true
    RevokedResponse:
      type: object
      required: [revoked]
      properties:
        revoked:
          type: boolean
          example: true
...